requirements.txt
Flask
uWSGI
Flask-Login
Flask-SQLAlchemy
docker-compose
requests
ruamel.yaml
rq
rq-dashboard
